move requests to an index file
